requirements.txt
Flask
gunicorn
torch
torchvision
Pillow
werkzeug>=2.0.0
